% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_compact_pipe.R
\name{gt_compact_pipe}
\alias{gt_compact_pipe}
\title{Compact pipe to gt table
also makes font roboto..}
\usage{
gt_compact_pipe(data)
}
\arguments{
\item{data}{}
}
\value{
gt
}
\description{
Compact pipe to gt table
also makes font roboto..
}
