% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma.R
\name{run_limma}
\alias{run_limma}
\title{Convience wrapper for actually running limma, not exported.}
\usage{
run_limma(exprs_values, design, genelist)
}
\arguments{
\item{exprs_values}{temp}

\item{design}{temp}

\item{genelist}{temp}
}
\value{
tibble
}
\description{
Convience wrapper for actually running limma, not exported.
}
