% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify_colnames.R
\name{prettify_colnames}
\alias{prettify_colnames}
\title{Prettifies colnames}
\usage{
prettify_colnames(df, prefix = "pr_")
}
\description{
This function prettifies colnames for quick plotting/tables
removes _ and makes space instead, makes titlecase and removes
an optional prefix.
}
\details{
@param df input dataframe
@prefix prefix to remove is set to remove processed 'pr_'

@export
@examples
prettify_colnames(df, prefix = 'hej)
}
