% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Plot volcano}
\usage{
volcano_plot(
  dataframe,
  adjusted = TRUE,
  nudge_y = 0.5,
  legend.position = "bottom_left"
)
}
\arguments{
\item{dataframe}{results form limma (or equivalent with col log_fc and p_value/adj_p_value and label)}

\item{adjusted}{adjusted or unadjusted p values}

\item{nudge_y}{if labels should be nudged.}

\item{legend.position}{where the legend should be}
}
\value{
ggplot2.
}
\description{
Takes a df which is output from limma (should have cols log_fc and p_value)
}
\examples{
volcanoplot(limma_results)
}
